Move command tests to OpenKh.Tests.Commands project
