+++ OpenKh.Tests.Commands/res/mapgen/diagnoal/diagnoal.coct.dump
@@ -1,0 +1,9 @@
+[CollisionMeshGroup] BoundingBox=((-1001, -1, -940), (1001, 1074, 940))
+ [CollisionMeshGroup] BoundingBox=((-1001, -1, -940), (1001, 1074, 940))
+  [CollisionMesh] BoundingBox=((-1001, -1, -940), (1001, 1074, 940)); v10=0; v12=0
+   [Collision] BoundingBox=((-1001, -1, -851), (1001, 895, 940)); Plane={Normal:<2.7287822E-08, -0.89449215, 0.44708368> D:379.86658}; v00=0; SurfaceFlags=000003F1; Vertices=<-1000.00006, -0.064697266, -849.7837, 1>, <999.99994, -0.064697266, -849.7838, 1>, <1000.00006, 894.10266, 939.20044, 1>, <-999.99994, 894.10266, 939.20056, 1>
+   [Collision] BoundingBox=((-1001, -1, -940), (-999, 1074, 940)); Plane={Normal:<-0.99999994, 2.7287824E-08, 5.4595464E-08> D:-999.99994}; v00=0; SurfaceFlags=000003F1; Vertices=<-1000.00006, 178.83374, -939.20044, 1>, <-1000.00006, -0.064697266, -849.7837, 1>, <-999.99994, 894.10266, 939.20056, 1>, <-999.99994, 1073.0011, 849.7838, 1>
+   [Collision] BoundingBox=((-1001, 893, 849), (1001, 1074, 940)); Plane={Normal:<5.371093E-08, 0.4470837, 0.8944921> D:-1239.8462}; v00=0; SurfaceFlags=000003F1; Vertices=<-999.99994, 1073.0011, 849.7838, 1>, <-999.99994, 894.10266, 939.20056, 1>, <1000.00006, 894.10266, 939.20044, 1>, <1000.00006, 1073.0011, 849.7837, 1>
+   [Collision] BoundingBox=((999, -1, -940), (1001, 1074, 940)); Plane={Normal:<0.99999994, -2.7287824E-08, -5.4595464E-08> D:-999.99994}; v00=0; SurfaceFlags=000003F1; Vertices=<1000.00006, 1073.0011, 849.7837, 1>, <1000.00006, 894.10266, 939.20044, 1>, <999.99994, -0.064697266, -849.7838, 1>, <999.99994, 178.83374, -939.20056, 1>
+   [Collision] BoundingBox=((-1001, -1, -940), (1001, 180, -849)); Plane={Normal:<-5.371093E-08, -0.4470837, -0.8944921> D:-760.1538}; v00=0; SurfaceFlags=000003F1; Vertices=<999.99994, 178.83374, -939.20056, 1>, <999.99994, -0.064697266, -849.7838, 1>, <-1000.00006, -0.064697266, -849.7837, 1>, <-1000.00006, 178.83374, -939.20044, 1>
+   [Collision] BoundingBox=((-1001, 178, -940), (1001, 1074, 851)); Plane={Normal:<-2.7287822E-08, 0.89449215, -0.44708368> D:-579.8666}; v00=0; SurfaceFlags=000003F1; Vertices=<999.99994, 178.83374, -939.20056, 1>, <-1000.00006, 178.83374, -939.20044, 1>, <-999.99994, 1073.0011, 849.7838, 1>, <1000.00006, 1073.0011, 849.7837, 1>
