+++ OpenKh.Tests.Commands/res/mapgen/diagnoal/diagnoal.doct.dump
@@ -1,0 +1,3 @@
+[Entry1] Unk=0; BoundingBox=Minimum: <-1001, -1, -940> Maximum: <1001, 1074, 940>
+ [Entry1] Unk=0; BoundingBox=Minimum: <-1001, -1, -940> Maximum: <1001, 1074, 940>
+  [Entry2] Flags=0; BoundingBox=Minimum: <-1001, -1, -940> Maximum: <1001, 1074, 940>
