+++ OpenKh.Tests.Commands/res/mapgen/diagnoal/diagnoal.fbx
