+++ OpenKh.Tests.Commands/res/mapgen/multi-4bpp/multi-4bpp.fbx
