+++ OpenKh.Tests.Commands/res/mapgen/multi-8bpp/multi-8bpp.fbx
