+++ OpenKh.Tests.Commands/res/mapgen/terrain/terrain.doct.dump
@@ -1,0 +1,3 @@
+[Entry1] Unk=0; BoundingBox=Minimum: <-835, -283, -835> Maximum: <835, 266, 835>
+ [Entry1] Unk=0; BoundingBox=Minimum: <-835, -283, -835> Maximum: <835, 266, 835>
+  [Entry2] Flags=0; BoundingBox=Minimum: <-835, -283, -835> Maximum: <835, 266, 835>
