+++ OpenKh.Tests.Commands/res/mapgen/terrain/terrain.fbx
