+++ OpenKh.Tests/OpenKh.Tests.csproj
@@ -166,42 +166,6 @@
     <None Update="kh2\res\map.spawnscript">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="kh2\res\mapgen\diagnoal\diagnoal.coct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\diagnoal\diagnoal.doct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\diagnoal\diagnoal.fbx">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\multi-4bpp\multi-4bpp.coct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\multi-4bpp\multi-4bpp.doct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\multi-4bpp\multi-4bpp.fbx">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\multi-8bpp\multi-8bpp.coct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\multi-8bpp\multi-8bpp.doct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\multi-8bpp\multi-8bpp.fbx">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\terrain\terrain.coct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\terrain\terrain.doct.dump">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="kh2\res\mapgen\terrain\terrain.fbx">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
     <None Update="kh2\res\model_texture1.tex">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
